requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
tensorflow==2.17.0
numpy==1.23.5
pillow==10.4.0
scikit-learn==1.5.1
matplotlib==3.9.1.post1
typing-extensions==4.12.2
python-multipart==0.0.9
gdown==5.2.0
requests>=2.25.1
firebase-admin==6.4.0